app-with-david/clockify/requirements.txt
Django~=4.1.4
djangorestframework~=3.14.0
djoser~=2.1.0